AbstractArray base & VectorAny (no  types in constrctor)
